Simplified module structure.
